Extracted out the appletv stuff to appletv.py
